test-server/requirements.txt
PyGithub==1.35
PyJWT==1.5.3
